fix(lens): move formatSend into render.ts with the other kaval-tui formatters
formatSend is a peer of formatList/shortId, so it lives in render.ts now; send.ts owns only encode/plan and no longer imports shortId. Its tests move to render.test.ts.

Agreed by the lowy ⇄ hickey lens debate (finding hickey-4, raised by hickey). Not pushed or merged.
